feat: spread logic of the contract among modules
