+++ contract/src/lib.rs
@@ -1,43 +1,33 @@
 use aurora_launchpad_types::config::{
-    DepositToken, DistributionProportions, LaunchpadConfig, LaunchpadStatus, Mechanics, TokenId,
+    DepositToken, DistributionProportions, LaunchpadConfig, LaunchpadStatus, Mechanics,
     VestingSchedule,
 };
-use aurora_launchpad_types::{
-    DistributionDirection, IntentAccount, InvestmentAmount, WithdrawDirection,
-};
-use near_plugins::{
-    AccessControlRole, AccessControllable, Pausable, Upgradable, access_control,
-    access_control_any, pause,
-};
+use aurora_launchpad_types::{IntentAccount, InvestmentAmount, WithdrawDirection};
+use near_plugins::{AccessControlRole, AccessControllable, Pausable, Upgradable, access_control};
 use near_sdk::borsh::BorshDeserialize;
 use near_sdk::json_types::U128;
-use near_sdk::serde_json::json;
 use near_sdk::store::{LazyOption, LookupMap};
-use near_sdk::{
-    AccountId, Gas, NearToken, PanicOnDefault, Promise, PromiseOrValue, PromiseResult,
-    assert_one_yocto, env, near, require,
-};
+use near_sdk::{AccountId, Gas, NearToken, PanicOnDefault, env, near};
 
 use crate::mechanics::claim::available_for_claim;
 use crate::storage_key::StorageKey;
-use crate::traits::{ext_ft, ext_mt};
-use crate::utils::parse_accounts;
 
+mod claim;
+mod deposit;
+mod distribute;
+mod lock;
 mod mechanics;
 mod storage_key;
 #[cfg(test)]
 mod tests;
 mod traits;
 mod utils;
+mod withdraw;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 const GAS_FOR_FT_TRANSFER_CALL: Gas = Gas::from_tgas(35);
 const GAS_FOR_FT_TRANSFER: Gas = Gas::from_tgas(3);
-const GAS_FOR_FINISH_CLAIM: Gas = Gas::from_tgas(2);
-const GAS_FOR_FINISH_DISTRIBUTION: Gas = Gas::from_tgas(1);
-const GAS_FOR_FINISH_WITHDRAW: Gas = Gas::from_tgas(1);
-
 const ONE_YOCTO: NearToken = NearToken::from_yoctonear(1);
 
 #[derive(AccessControlRole, Clone, Copy)]
@@ -277,432 +267,6 @@ impl AuroraLaunchpadContract {
     #[must_use]
     pub const fn get_version() -> &'static str {
         VERSION
-    }
-
-    /// Sets the status of the contract is `Locked`.
-    #[access_control_any(roles(Role::Admin))]
-    pub fn lock(&mut self) {
-        let status = self.get_status();
-        require!(
-            status == LaunchpadStatus::NotStarted || status == LaunchpadStatus::Ongoing,
-            "The contract is not started nor ongoing"
-        );
-
-        near_sdk::log!("The contract is locked");
-
-        self.is_locked = true;
-    }
-
-    /// Unsets the `Locked` status from the contract.
-    #[access_control_any(roles(Role::Admin))]
-    pub fn unlock(&mut self) {
-        require!(
-            self.get_status() == LaunchpadStatus::Locked,
-            "The contract is not locked"
-        );
-
-        near_sdk::log!("The contract is unlocked");
-
-        self.is_locked = false;
-    }
-
-    #[pause]
-    #[payable]
-    pub fn claim(&mut self, withdraw_direction: WithdrawDirection) -> Promise {
-        assert_one_yocto();
-        require!(
-            self.is_success(),
-            "Claim can be called only if the launchpad finishes with success status"
-        );
-
-        let predecessor_account_id = env::predecessor_account_id();
-
-        let intents_account_id =
-            self.get_intents_account_id(&withdraw_direction, &predecessor_account_id);
-
-        let Some(investment) = self.investments.get_mut(&intents_account_id) else {
-            env::panic_str("No deposits found for the intent account");
-        };
-        // available_for_claim - claimed
-        let assets_amount = match available_for_claim(
-            investment,
-            self.total_sold_tokens,
-            &self.config,
-            env::block_timestamp(),
-        ) {
-            Ok(amount) => amount.saturating_sub(investment.claimed),
-            Err(err) => env::panic_str(&format!("Claim failed: {err}")),
-        };
-
-        match withdraw_direction {
-            WithdrawDirection::Intents(_) => ext_ft::ext(self.config.sale_token_account_id.clone())
-                .with_attached_deposit(ONE_YOCTO)
-                .with_static_gas(GAS_FOR_FT_TRANSFER_CALL)
-                .ft_transfer_call(
-                    self.config.intents_account_id.clone(),
-                    assets_amount.into(),
-                    intents_account_id.as_ref().to_string(),
-                    None,
-                ),
-            WithdrawDirection::Near => ext_ft::ext(self.config.sale_token_account_id.clone())
-                .with_attached_deposit(ONE_YOCTO)
-                .with_static_gas(GAS_FOR_FT_TRANSFER)
-                .ft_transfer(predecessor_account_id, assets_amount.into(), None),
-        }
-        .then(
-            Self::ext(env::current_account_id())
-                .with_static_gas(GAS_FOR_FINISH_CLAIM)
-                .finish_claim(&intents_account_id, assets_amount),
-        )
-    }
-
-    #[private]
-    pub fn finish_claim(&mut self, intent_account_id: &IntentAccount, assets_amount: u128) {
-        require!(
-            env::promise_results_count() == 1,
-            "Expected one promise result"
-        );
-
-        match env::promise_result(0) {
-            PromiseResult::Successful(_) => {
-                let Some(investment) = self.investments.get_mut(intent_account_id) else {
-                    env::panic_str("No deposits found for the intent account");
-                };
-                // Increase claimed assets
-                investment.claimed = investment.claimed.saturating_add(assets_amount);
-            }
-            PromiseResult::Failed => {
-                env::panic_str("Claim transfer failed");
-            }
-        }
-    }
-
-    #[pause]
-    #[payable]
-    pub fn withdraw(&mut self, amount: U128, withdraw_direction: WithdrawDirection) -> Promise {
-        assert_one_yocto();
-        let status = self.get_status();
-        let is_price_discovery_ongoing = matches!(self.config.mechanics, Mechanics::PriceDiscovery)
-            && matches!(status, LaunchpadStatus::Ongoing);
-
-        require!(
-            !(is_price_discovery_ongoing
-                && matches!(withdraw_direction, WithdrawDirection::Intents(_))),
-            "Withdraw is not allowed to Intents in PriceDiscovery mechanics and Ongoing status"
-        );
-
-        let is_withdrawal_allowed = is_price_discovery_ongoing
-            || matches!(status, LaunchpadStatus::Failed)
-            || matches!(status, LaunchpadStatus::Locked);
-        require!(is_withdrawal_allowed, "Withdraw is not allowed");
-
-        let predecessor_account_id = env::predecessor_account_id();
-        let intents_account_id =
-            self.get_intents_account_id(&withdraw_direction, &predecessor_account_id);
-
-        let Some(investment) = self.investments.get(&intents_account_id) else {
-            env::panic_str("No deposits found for the intent account");
-        };
-
-        mechanics::withdraw::validate_amount(investment, amount.0, &self.config)
-            .unwrap_or_else(|err| env::panic_str(err));
-
-        match withdraw_direction {
-            WithdrawDirection::Intents(_) => self.withdraw_to_intents(&intents_account_id, amount),
-            WithdrawDirection::Near => self.withdraw_to_near(predecessor_account_id, amount),
-        }
-        .then(
-            Self::ext(env::current_account_id())
-                .with_static_gas(GAS_FOR_FINISH_WITHDRAW)
-                .finish_withdraw(&intents_account_id, amount.0, env::block_timestamp()),
-        )
-    }
-
-    #[private]
-    pub fn finish_withdraw(&mut self, intent_account_id: &IntentAccount, amount: u128, time: u64) {
-        require!(
-            env::promise_results_count() == 1,
-            "Expected one promise result"
-        );
-
-        match env::promise_result(0) {
-            PromiseResult::Successful(_) => {
-                let Some(investment) = self.investments.get_mut(intent_account_id) else {
-                    env::panic_str("No deposits found for the intent account");
-                };
-
-                mechanics::withdraw::post_withdraw(
-                    investment,
-                    amount,
-                    &mut self.total_deposited,
-                    &mut self.total_sold_tokens,
-                    &self.config,
-                    time,
-                )
-                .unwrap_or_else(|err| env::panic_str(&format!("Withdraw failed: {err}")));
-            }
-            PromiseResult::Failed => {
-                env::panic_str("Withdraw transfer failed");
-            }
-        }
-    }
-
-    #[pause]
-    #[payable]
-    pub fn distribute_tokens(&mut self, distribution_direction: &DistributionDirection) -> Promise {
-        require!(
-            self.is_success(),
-            "Distribution can be called only if the launchpad finishes with success status"
-        );
-        require!(!self.is_distributed, "Tokens have been already distributed");
-
-        match distribution_direction {
-            DistributionDirection::Intents => self.distribute_to_intents(),
-            DistributionDirection::Near => self.distribute_to_near(),
-        }
-        .then(
-            Self::ext(env::current_account_id())
-                .with_static_gas(GAS_FOR_FINISH_DISTRIBUTION)
-                .finish_distribution(),
-        )
-    }
-
-    #[private]
-    pub fn finish_distribution(&mut self) {
-        require!(
-            env::promise_results_count() > 0,
-            "Expected at least one promise result"
-        );
-
-        match env::promise_result(0) {
-            PromiseResult::Successful(_) => self.is_distributed = true,
-            PromiseResult::Failed => env::panic_str("Distribution failed"),
-        }
-    }
-
-    #[pause]
-    pub fn ft_on_transfer(
-        &mut self,
-        sender_id: AccountId,
-        amount: U128,
-        msg: String,
-        memo: Option<String>,
-    ) -> PromiseOrValue<U128> {
-        let _ = (sender_id, memo);
-        let token_account_id = env::predecessor_account_id();
-
-        if token_account_id == self.config.sale_token_account_id {
-            self.init_contract(amount)
-        } else if self.is_nep141_deposit_token(&token_account_id) {
-            self.handle_deposit(amount, &msg)
-        } else {
-            env::panic_str("Unsupported NEP-141 token");
-        }
-    }
-
-    #[pause]
-    pub fn mt_on_transfer(
-        &mut self,
-        sender_id: AccountId,
-        previous_owner_ids: Vec<AccountId>,
-        token_ids: Vec<TokenId>,
-        amounts: Vec<U128>,
-        msg: String,
-    ) -> PromiseOrValue<Vec<U128>> {
-        let _ = (sender_id, previous_owner_ids);
-        require!(
-            self.is_nep245_deposit_token(&env::predecessor_account_id(), &token_ids),
-            "Wrong NEP-245 deposit token"
-        );
-
-        match self.handle_deposit(amounts[0], &msg) {
-            PromiseOrValue::Promise(promise) => PromiseOrValue::Promise(promise),
-            PromiseOrValue::Value(value) => PromiseOrValue::Value(vec![value]),
-        }
-    }
-
-    fn init_contract(&mut self, amount: U128) -> PromiseOrValue<U128> {
-        if self.is_sale_token_set {
-            env::panic_str("The contract is already initialized");
-        }
-
-        require!(
-            amount == self.config.total_sale_amount,
-            "Wrong total sale amount"
-        );
-
-        near_sdk::log!("The contract has been initialized successfully");
-
-        self.is_sale_token_set = true;
-        PromiseOrValue::Value(0.into())
-    }
-
-    fn distribute_to_intents(&self) -> Promise {
-        let promise_res = ext_ft::ext(self.config.sale_token_account_id.clone())
-            .with_attached_deposit(ONE_YOCTO)
-            .with_static_gas(GAS_FOR_FT_TRANSFER_CALL)
-            .ft_transfer_call(
-                self.config.intents_account_id.clone(),
-                self.config.distribution_proportions.solver_allocation,
-                self.config
-                    .distribution_proportions
-                    .solver_account_id
-                    .as_ref()
-                    .to_string(),
-                None,
-            );
-
-        self.config
-            .distribution_proportions
-            .stakeholder_proportions
-            .iter()
-            .fold(promise_res, |promise, proportion| {
-                promise.function_call(
-                    "ft_transfer_call".to_string(),
-                    json!({
-                        "receiver_id": self.config.intents_account_id.clone(),
-                        "amount": proportion.allocation,
-                        "msg": proportion.account.as_ref().to_string(),
-                    })
-                    .to_string()
-                    .into_bytes(),
-                    ONE_YOCTO,
-                    GAS_FOR_FT_TRANSFER_CALL,
-                )
-            })
-    }
-
-    fn distribute_to_near(&self) -> Promise {
-        let promise = ext_ft::ext(self.config.sale_token_account_id.clone())
-            .with_attached_deposit(ONE_YOCTO)
-            .with_static_gas(GAS_FOR_FT_TRANSFER)
-            .ft_transfer(
-                self.config
-                    .distribution_proportions
-                    .solver_account_id
-                    .clone()
-                    .try_into()
-                    .unwrap(),
-                self.config.distribution_proportions.solver_allocation,
-                None,
-            );
-
-        self.config
-            .distribution_proportions
-            .stakeholder_proportions
-            .iter()
-            .fold(promise, |promise, proportion| {
-                let receiver_id: AccountId = proportion
-                    .account
-                    .clone()
-                    .try_into()
-                    .unwrap_or_else(|e| env::panic_str(e));
-                promise.function_call(
-                    "ft_transfer".to_string(),
-                    json!({
-                        "receiver_id": receiver_id,
-                        "amount": proportion.allocation,
-                    })
-                    .to_string()
-                    .into_bytes(),
-                    ONE_YOCTO,
-                    GAS_FOR_FT_TRANSFER,
-                )
-            })
-    }
-
-    fn withdraw_to_intents(&self, intents_account: &IntentAccount, amount: U128) -> Promise {
-        match &self.config.deposit_token {
-            DepositToken::Nep141(account_id) => ext_ft::ext(account_id.clone())
-                .with_attached_deposit(ONE_YOCTO)
-                .with_static_gas(GAS_FOR_FT_TRANSFER_CALL)
-                .ft_transfer_call(
-                    self.config.intents_account_id.clone(),
-                    amount,
-                    intents_account.as_ref().to_string(),
-                    None,
-                ),
-            DepositToken::Nep245((account_id, token_id)) => ext_mt::ext(account_id.clone())
-                .with_attached_deposit(ONE_YOCTO)
-                .with_static_gas(GAS_FOR_FT_TRANSFER_CALL)
-                .mt_transfer_call(
-                    self.config.intents_account_id.clone(),
-                    token_id.clone(),
-                    amount,
-                    None,
-                    None,
-                    intents_account.as_ref().to_string(),
-                ),
-        }
-    }
-
-    fn withdraw_to_near(&self, receiver_id: AccountId, amount: U128) -> Promise {
-        match &self.config.deposit_token {
-            DepositToken::Nep141(account_id) => ext_ft::ext(account_id.clone())
-                .with_attached_deposit(ONE_YOCTO)
-                .with_static_gas(GAS_FOR_FT_TRANSFER)
-                .ft_transfer(receiver_id, amount, None),
-            DepositToken::Nep245((account_id, token_id)) => ext_mt::ext(account_id.clone())
-                .with_attached_deposit(ONE_YOCTO)
-                .with_static_gas(GAS_FOR_FT_TRANSFER)
-                .mt_transfer(receiver_id, token_id.clone(), amount, None, None),
-        }
-    }
-
-    fn handle_deposit(&mut self, amount: U128, msg: &str) -> PromiseOrValue<U128> {
-        require!(self.is_ongoing(), "Launchpad is not ongoing");
-        // Get NEAR and IntentAccount from the message
-        let (near_account_id, intent_account_id) =
-            parse_accounts(msg).unwrap_or_else(|err| env::panic_str(err));
-
-        // Insert IntentAccount to the accounts map if near_account_id was provided
-        // and it doesn't exist
-        if let Some(near_account_id) = near_account_id {
-            self.accounts
-                .entry(near_account_id)
-                .or_insert_with(|| intent_account_id.clone());
-        }
-
-        near_sdk::log!("Depositing amount: {} for: {intent_account_id}", amount.0);
-
-        let investments = self
-            .investments
-            .entry(intent_account_id)
-            .or_insert_with(|| {
-                self.participants_count += 1;
-                InvestmentAmount::default()
-            });
-
-        let deposit_result = mechanics::deposit::deposit(
-            investments,
-            amount.0,
-            &mut self.total_deposited,
-            &mut self.total_sold_tokens,
-            &self.config,
-            env::block_timestamp(),
-        );
-        let remain = match deposit_result {
-            Ok(val) => val,
-            Err(err) => env::panic_str(&format!("Deposit failed: {err}")),
-        };
-
-        PromiseOrValue::Value(remain.into())
-    }
-
-    fn is_nep141_deposit_token(&self, predecessor_account_id: &AccountId) -> bool {
-        matches!(&self.config.deposit_token, DepositToken::Nep141(account_id) if account_id == predecessor_account_id)
-    }
-
-    fn is_nep245_deposit_token(
-        &self,
-        predecessor_account_id: &AccountId,
-        token_ids: &[TokenId],
-    ) -> bool {
-        require!(
-            token_ids.len() == 1,
-            "Only one token_id is allowed for deposit"
-        );
-        matches!(&self.config.deposit_token, DepositToken::Nep245((account_id, token_id)) if account_id == predecessor_account_id && token_id == &token_ids[0])
     }
 
     fn get_intents_account_id(
