+++ tests/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "tests"
-version = "0.1.0"
-edition = "2024"
+name = "aurora-launchpad-tests"
+version.workspace = true
+edition.workspace = true
+publish.workspace = true
 
 [dependencies]
 anyhow.workspace = true
